// src/api/models/StoreReq.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StoreReq = {
    storeId: number;
    storeName: string;
    serviceIndustry: StoreReq.serviceIndustry;
    gu?: string;
    dong: string;
    openingDate: string;
    brandCode?: number;
    franchise?: boolean;
};
export namespace StoreReq {
    export enum serviceIndustry {
        CS100001 = 'CS100001',
        CS100002 = 'CS100002',
        CS100003 = 'CS100003',
        CS100004 = 'CS100004',
        CS100005 = 'CS100005',
        CS100006 = 'CS100006',
        CS100007 = 'CS100007',
        CS100008 = 'CS100008',
        CS100009 = 'CS100009',
        CS100010 = 'CS100010',
        CS200001 = 'CS200001',
        CS200002 = 'CS200002',
        CS200003 = 'CS200003',
        CS200005 = 'CS200005',
        CS200006 = 'CS200006',
        CS200007 = 'CS200007',
        CS200008 = 'CS200008',
        CS200016 = 'CS200016',
        CS200017 = 'CS200017',
        CS200019 = 'CS200019',
        CS200024 = 'CS200024',
        CS200025 = 'CS200025',
        CS200026 = 'CS200026',
        CS200028 = 'CS200028',
        CS200029 = 'CS200029',
        CS200030 = 'CS200030',
        CS200031 = 'CS200031',
        CS200032 = 'CS200032',
        CS200033 = 'CS200033',
        CS200034 = 'CS200034',
        CS200036 = 'CS200036',
        CS200037 = 'CS200037',
        CS300001 = 'CS300001',
        CS300002 = 'CS300002',
        CS300003 = 'CS300003',
        CS300004 = 'CS300004',
        CS300006 = 'CS300006',
        CS300007 = 'CS300007',
        CS300008 = 'CS300008',
        CS300009 = 'CS300009',
        CS300010 = 'CS300010',
        CS300011 = 'CS300011',
        CS300014 = 'CS300014',
        CS300015 = 'CS300015',
        CS300016 = 'CS300016',
        CS300017 = 'CS300017',
        CS300018 = 'CS300018',
        CS300019 = 'CS300019',
        CS300020 = 'CS300020',
        CS300021 = 'CS300021',
        CS300022 = 'CS300022',
        CS300024 = 'CS300024',
        CS300025 = 'CS300025',
        CS300026 = 'CS300026',
        CS300027 = 'CS300027',
        CS300028 = 'CS300028',
        CS300029 = 'CS300029',
        CS300031 = 'CS300031',
        CS300032 = 'CS300032',
        CS300033 = 'CS300033',
        CS300035 = 'CS300035',
        CS300036 = 'CS300036',
        CS300043 = 'CS300043',
    }
}